ptcloud: address review (naming, BPA perf/robustness, shared helper, test/perf consolidation)
