+++ modules/ptcloud/src/normal_estimation.cpp
@@ -66,11 +66,7 @@ void orientNormalsConsistent(InputArray inputCloud, InputOutputArray normals, in
 
     std::vector<Vec3f> n(N);
     for (int i = 0; i < N; i++)
-    {
-        Vec3f v = nmf.at<Vec3f>(i);
-        float len = (float)cv::norm(v);
-        n[i] = (len > 1e-12f) ? v * (1.f / len) : Vec3f(0.f, 0.f, 1.f);   // guard degenerate normals
-    }
+        n[i] = safeNormalize(nmf.at<Vec3f>(i));
 
     // Riemannian graph over kNN, weight 1 - |n_i . n_j| (Hoppe et al. 1992).
     Mat pts = Mat(points).reshape(1, N);
@@ -79,15 +75,27 @@ void orientNormalsConsistent(InputArray inputCloud, InputOutputArray normals, in
     Mat nnIdx, nnDist;
     index.knnSearch(pts, nnIdx, nnDist, kk);
 
+    // Build the kNN edges in parallel: each point writes to its own fixed slots, then compact.
+    std::vector<MSTEdge> slots((size_t)N * (kk - 1));
+    std::vector<char> valid((size_t)N * (kk - 1), 0);
+    parallel_for_(Range(0, N), [&](const Range& range)
+    {
+        for (int i = range.start; i < range.end; i++)
+            for (int c = 1; c < kk; c++)              // skip self at column 0
+            {
+                int j = nnIdx.at<int>(i, c);
+                if (j >= 0 && j != i)
+                {
+                    size_t s = (size_t)i * (kk - 1) + (c - 1);
+                    slots[s] = MSTEdge{ i, j, 1.0 - std::fabs((double)n[i].dot(n[j])) };
+                    valid[s] = 1;
+                }
+            }
+    });
     std::vector<MSTEdge> edges;
-    edges.reserve((size_t)N * kk);
-    for (int i = 0; i < N; i++)
-        for (int c = 1; c < kk; c++)                  // skip self at column 0
-        {
-            int j = nnIdx.at<int>(i, c);
-            if (j >= 0 && j != i)
-                edges.push_back(MSTEdge{ i, j, 1.0 - std::fabs((double)n[i].dot(n[j])) });
-        }
+    edges.reserve(slots.size());
+    for (size_t s = 0; s < slots.size(); s++)
+        if (valid[s]) edges.push_back(slots[s]);
 
     // Seed the +Z-extreme point outward; orientation propagates along the MST.
     int seed = 0;
