+++ modules/ptcloud/src/ptcloud_utils.hpp
@@ -44,6 +44,13 @@ static inline void toPointVec(InputArray inputCloud, std::vector<Point3f>& point
     points.assign(mf.begin<Point3f>(), mf.end<Point3f>());
 }
 
+// Normalize a vector; fall back to a fixed unit vector for (near-)zero length.
+static inline Vec3f safeNormalize(const Vec3f& v)
+{
+    float len = (float)norm(v);
+    return (len > 1e-12f) ? v * (1.f / len) : Vec3f(0.f, 0.f, 1.f);
+}
+
 } // namespace cv
 
 #endif // OPENCV_PTCLOUD_UTILS_HPP
